Minimal Json Return Routes and Authorization Issues
I noticed that the user password was being exposed through the json returns. I altered my routes to prevent that from happening. Also, I noticed that I made the mistake of returning authorizations through the header even when an error occurred. These are minimal changes, but the whole process of debugging if something goes wrong is too stressful with the time we have left.  

Routes where password is exposed: 
-Login
-Logout
-Student Signup
-Tutor Signup

Routes where authorization is returned when it shouldn't be:
-Student Signup
-Tutor Signup
-Request Session
-Accept Session

Routes where I forgot to even return authorization in header:
-Rate Session

Note for me: Request sessions allows you to create a session with an invalid course.